Use pytest for tests
