+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+log_cli = true
+log_level = INFO
